allow user to set scope
